+++ flow-report/hooks.ts
@@ -1,0 +1,13 @@
+/**
+ * @license Copyright 2021 The Lighthouse Authors. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+ */
+
+// TODO(FR-COMPAT): Use router to handle steps.
+export function useCurrentStep() {
+  const searchParams = new URLSearchParams(location.search);
+  const step = searchParams.get('step');
+  if (step === null) return null;
+  return Number(step);
+}
